Rename: 패키지를 세분화한다.
