--- src/main/java/org/prgms/springbooturlshortener/domain/shorturl/service/UrlTransformer.java
@@ -1,5 +1,0 @@
-package org.prgms.springbooturlshortener.domain.shorturl.service;
-
-public interface UrlTransformer {
-    String generateUrl(int number);
-}
