+++ src/main/java/org/prgms/springbooturlshortener/domain/shorturl/service/transformer/UrlTransformer.java
@@ -1,0 +1,5 @@
+package org.prgms.springbooturlshortener.domain.shorturl.service.transformer;
+
+public interface UrlTransformer {
+    String generateUrl(int number);
+}
